refactor some things
