+++ src/main/java/io/ioxcorp/ioxbox/Main.java
@@ -2,8 +2,6 @@ package io.ioxcorp.ioxbox;
 
 import io.ioxcorp.ioxbox.data.format.Box;
 import io.ioxcorp.ioxbox.data.json.JacksonYeehawHelper;
-import io.ioxcorp.ioxbox.listeners.ConfirmationDialogListener;
-import io.ioxcorp.ioxbox.listeners.ConfirmationGetter;
 import io.ioxcorp.ioxbox.listeners.ConfirmationGetterListener;
 import io.ioxcorp.ioxbox.listeners.Listener;
 import net.dv8tion.jda.api.JDA;
