+++ src/main/java/io/ioxcorp/ioxbox/listeners/HandleDelete.java
@@ -2,6 +2,8 @@ package io.ioxcorp.ioxbox.listeners;
 
 import io.ioxcorp.ioxbox.Main;
 import io.ioxcorp.ioxbox.data.format.CustomUser;
+import io.ioxcorp.ioxbox.data.format.WhatAmIDoing;
+import io.ioxcorp.ioxbox.listeners.util.ConfirmationGetter;
 
 public class HandleDelete implements Runnable {
     private final CustomUser user;
